Last Version Updated

+++ Shipping.BL/DTOs/Weight/AddWeightDTO.cs
@@ -9,9 +9,7 @@ namespace Shipping.BL.DTOs.Weight
     public class AddWeightDTO
     {
         public int Id { get; set; }
-        public decimal DefualtWeight {  get; set; } //Not in SRS only Default Price na extra price per Kilo 
+        public int DefualtWeight {  get; set; } //Not in SRS only Default Price na extra price per Kilo 
         public decimal ExtraPricePerKilo {  get; set; }
-
-
     }
 }
